+++ swaps/get_paid_service.js
@@ -1,5 +1,6 @@
 const asyncAuto = require('async/auto');
 const {encode} = require('cbor');
+const {genericSwapService} = require('goldengate');
 const {getNetwork} = require('ln-sync');
 const {getSwapMacaroon} = require('goldengate');
 const {lightningLabsSwapService} = require('goldengate');
@@ -10,12 +11,18 @@ const {swapUserId} = require('goldengate');
 const decodeSwapApiKey = require('./decode_swap_api_key');
 const {probeDestination} = require('./../network');
 
+const bitcoinTestnetNetwork = 'btctestnet';
+const bufferFromBase64 = base64 => Buffer.from(base64, 'base64');
+const bufferFromHex = hex => Buffer.from(hex, 'hex');
 const encodeCbor = json => encode(json).toString('hex');
+const httpMatch = /^https?:\/\//;
 const maxRoutingFee = 100;
+const testnetSocket = 'https://balanceofsatoshis.com:11010';
 
 /** Get a paid swap service object
 
   {
+    fetch: <Fetch Function>
     lnd: <Authenticated LND gRPC API Object>
     logger: <Winston Logger Object>
     [socket]: <Custom Backing Service Socket String>
@@ -32,11 +39,15 @@ const maxRoutingFee = 100;
     token: <Authentication Token Hex String>
   }
 */
-module.exports = ({lnd, logger, socket, token}, cbk) => {
+module.exports = ({fetch, lnd, logger, socket, token}, cbk) => {
   return new Promise((resolve, reject) => {
     return asyncAuto({
       // Check arguments
       validate: cbk => {
+        if (!fetch) {
+          return cbk([400, 'ExpectedFetchFunctionToGetPaidService']);
+        }
+
         if (!lnd) {
           return cbk([400, 'ExpectedLndToGetPaidService']);
         }
@@ -61,11 +72,36 @@ module.exports = ({lnd, logger, socket, token}, cbk) => {
       // Get network
       getNetwork: ['validate', ({}, cbk) => getNetwork({lnd}, cbk)],
 
+      // Make a service object for the remote swap service
+      remote: ['getNetwork', ({getNetwork}, cbk) => {
+        // Exit early when using a generic service
+        if (!!socket && httpMatch.test(socket)) {
+          try {
+            return cbk(null, genericSwapService({fetch, socket}));
+          } catch (err) {
+            return cbk([500, 'UnexpectedErrorInitiatingSwapService', {err}]);
+          }
+        }
+
+        const {network} = getNetwork;
+
+        // Exit early and use a generic service on testnet
+        if (!socket && network === bitcoinTestnetNetwork) {
+          return cbk(null, genericSwapService({fetch, socket: testnetSocket}));
+        }
+
+        try {
+          return cbk(null, lightningLabsSwapService({network, socket}));
+        } catch (err) {
+          return cbk([500, 'UnexpectedErrorInitiatingSwapService', {err}]);
+        }
+      }],
+
       // Get an unpaid swap macaroon
       getUnpaidMacaroon: [
         'decodeToken',
-        'getNetwork',
-        ({decodeToken, getNetwork}, cbk) =>
+        'remote',
+        ({decodeToken, remote}, cbk) =>
       {
         // Exit early when there is already a service token macaroon
         if (!!decodeToken) {
@@ -75,17 +111,7 @@ module.exports = ({lnd, logger, socket, token}, cbk) => {
           });
         }
 
-        const {network} = getNetwork;
-
-        try {
-          lightningLabsSwapService({network, socket});
-        } catch (err) {
-          return cbk([400, 'UnexpectedErrorInitiatingSwapService', {err}]);
-        }
-
-        const {service} = lightningLabsSwapService({network, socket});
-
-        return getSwapMacaroon({service}, cbk);
+        return getSwapMacaroon({service: remote.service}, cbk);
       }],
 
       // Pay for the macaroon
@@ -110,44 +136,32 @@ module.exports = ({lnd, logger, socket, token}, cbk) => {
         cbk);
       }],
 
-      // Get the paid service object
-      service: [
+      // Final service method details
+      paidService: [
         'decodeToken',
-        'getNetwork',
         'getUnpaidMacaroon',
         'payForMacaroon',
-        ({decodeToken, getNetwork, getUnpaidMacaroon, payForMacaroon}, cbk) =>
+        'remote',
+        ({decodeToken, getUnpaidMacaroon, payForMacaroon, remote}, cbk) =>
       {
         const {id} = getUnpaidMacaroon;
         const {macaroon} = getUnpaidMacaroon;
-        const {network} = getNetwork;
         const {paid} = payForMacaroon;
         const {preimage} = payForMacaroon;
+        const {service} = remote;
 
         if (!decodeToken && !paid) {
           return cbk([400, 'FailedToPurchasePaidServiceTokens']);
         }
 
-        try {
-          lightningLabsSwapService({macaroon, network, socket});
-        } catch (err) {
-          return cbk([400, 'FailedToFindSupportedSwapService', {err}]);
-        }
-
-        const {service} = lightningLabsSwapService({
-          macaroon,
-          network,
-          socket,
-        });
-
         const token = encodeCbor({
-          macaroon: Buffer.from(macaroon, 'base64'),
-          preimage: Buffer.from(preimage, 'hex'),
+          macaroon: bufferFromBase64(macaroon),
+          preimage: bufferFromHex(preimage),
         });
 
         return cbk(null, {id, macaroon, paid, preimage, service, token});
       }],
     },
-    returnResult({reject, resolve, of: 'service'}, cbk));
+    returnResult({reject, resolve, of: 'paidService'}, cbk));
   });
 };